fix(core): close note/derivation governance gaps per product doc v2
- retire POST /api/conversations/{id}/note/publish (governance bypass:
  skipped publishing mode, human confirmation, review and audit); notes
  publish via the shared chain commit-draft -> confirm -> publish /
  publication-requests
- note/sync dedup: unchanged message cursor is an idempotent no-op
  (status=unchanged) instead of minting identical versions; the upserted
  draft advances committed_revision so it is no longer perpetually dirty
- derivation source blocks now include note-container blocks (origin=note)
  alongside chat messages and chat-container blocks; derivation_sources
  records the per-block container; GET /api/derivations/{id} returns the
  per-block sources (with origin) for traceability
- context_policy is now effective: summary_only (200c/block digest),
  selected_only (2000c excerpts), full (source note markdown + excerpts)
  seed the derived conversation with an opening assistant message
- finalize reference/merge lineage materializes into asset.asset_relations
  (derived_from/continues_from, source=manual, derivation id in citation)
  inside the version's creating transaction, visible on
  GET /api/assets/{id}/relations; drop the write-only
  content.asset_relations (+asset_relation_blocks) via migration 0018
- fix pre-existing 500s surfaced by the new flow: staged draft relations
  query selected a non-existent dr.id; relation materialization failed on
  pgx type inference for the confidence CASE branch (::numeric)
- unify member conversation/note/derivation (+asset lineage) responses on
  the {"data":...,"request_id"} envelope; openapi now matches reality
  (note/sync and finalize are 201, NotePublishResult removed, Derivation
  carries sources); update yxt contracts and the frontend integration guide

+++ internal/asset/draft_service.go
@@ -502,7 +502,7 @@ func materializeVersionRelationsTx(ctx context.Context, tx pgx.Tx, material Vers
 				(organization_id, source_asset_version_id, target_asset_version_id,
 				 relation_type, source, confidence, citation, suggestion_id, created_by)
 			VALUES ($1::uuid, $2::uuid, $3::uuid, $4, $9,
-			        CASE WHEN $9 = 'ai' THEN $5 ELSE NULL END,
+			        CASE WHEN $9 = 'ai' THEN $5::numeric ELSE NULL END,
 			        $6::jsonb, NULLIF($7, '')::uuid, $8::uuid)
 			ON CONFLICT (source_asset_version_id, target_asset_version_id, relation_type) DO NOTHING
 			RETURNING id::text
